requirements.txt
ocrd-pyexiftool==0.2.0
PyYAML==5.1.2
wxPython>=4.0.0
